desc

+++ lib/tasks/custom.rake
@@ -61,6 +61,33 @@ namespace :pk do
     end
   end
 
+  task desc: :environment do
+
+    API = "http://pokeapi.co"
+    progress = ProgressBar.create(title: 'update description', starting_at: 0, total: 778)
+    pokemons = Pokemon.all
+    ActiveRecord::Base.transaction do
+      pokemons.each do |p|
+        if p.data['descriptions'][0]['resource_uri'].any?
+          desc = "No life, nothing to tell about him"
+        else
+          des = p.data['descriptions'][0]['resource_uri']
+          desc = HTTParty.get("#{API}#{des}")['description']
+        end
+        Pokemon.update(p.id,
+          description: esc
+        )
+
+        progress.increment
+      end
+    end
+  end
+if data['sprites'].any?
+            avatar = HTTParty.get("#{API}#{data['sprites'][0]['resource_uri']}")['image']
+            avatar = "#{API}#{avatar}"
+          else
+            avatar = "http://www.google.fr/imgres?imgurl=http%3A%2F%2Fstatic.planetminecraft.com%2Ffiles%2Fresource_media%2Fscreenshot%2F1318%2FQuestion_mark_alternate_5357730.jpg&imgrefurl=http%3A%2F%2Fwww.planetminecraft.com%2Fblog%2Fwant-a-minecraft-server-trailer%2F&h=640&w=494&tbnid=yFKpmq0TBHbJrM%3A&zoom=1&docid=nTQSmeMxkMn9VM&ei=9_ZYVO25I6zdsATXyIHABA&tbm=isch&iact=rc&uact=3&dur=179&page=3&start=74&ndsp=32&ved=0CLYCEK0DMFs"
+          end
   task reset: :environment do
     `rake db:drop:all`
     `rake db:create:all`
@@ -69,7 +96,7 @@ namespace :pk do
     `rake pk:vote`
   end
 
-  task reset_heroku: :environment do
+  task reset_heroku: :production do
     `heroku run rake db:drop:all`
     `heroku run rake db:create:all`
     `heroku run rake db:migrate`
